refine jsdoc and split config
